Move some classes to sub package
