+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/eventregistration/events/ui/category/TraceLocationCategoryFragment.kt
@@ -1,4 +1,4 @@
-package de.rki.coronawarnapp.eventregistration.events.ui
+package de.rki.coronawarnapp.eventregistration.events.ui.category
 
 import android.os.Bundle
 import android.view.View
